Moved month- and year-picker inside datepicker
Only added index exports. Nothing else changed yet.
