+++ libs/core/src/components/datepicker/index.ts
@@ -1,1 +1,3 @@
 export * from './datepicker'
+export * from './month-picker/month-picker'
+export * from './year-picker/year-picker'
