+++ libs/core/src/components/datepicker/month-picker/_mixins.scss
