+++ libs/core/src/components/datepicker/month-picker/month-picker.stories.ts
@@ -2,12 +2,12 @@ import { html } from 'lit'
 
 import type { Meta, StoryObj } from '@storybook/web-components'
 
-import { argTablePropsFor } from '../../../.storybook/argTableProps.ts'
+import { argTablePropsFor } from '../../../../.storybook/argTableProps.ts'
 
 import './index.ts'
-import '../popover/index.ts'
-import '../button/index.ts'
-import '../icon/icons/calendar.ts'
+import '../../popover/index.ts'
+import '../../button/index.ts'
+import '../../icon/icons/calendar.ts'
 
 /**
  * [Source code](https://github.com/seb-oss/green/tree/main/libs/core/src/components/month-picker)
