+++ libs/core/src/components/datepicker/month-picker/month-picker.styles.ts
