+++ libs/core/src/components/datepicker/month-picker/month-picker.trans.styles.scss
