+++ libs/core/src/components/datepicker/month-picker/month-picker.trans.styles.ts
@@ -1,4 +1,4 @@
-import { TransitionalStyles } from '../../utils/transitional-styles/transitional-styles'
+import { TransitionalStyles } from '../../../utils/transitional-styles/transitional-styles'
 import styles from './month-picker.trans.styles.scss?inline'
 
 export function register() {
