+++ libs/core/src/components/datepicker/year-picker/_mixins.scss
