+++ libs/core/src/components/datepicker/year-picker/year-picker.styles.ts
