+++ libs/core/src/components/datepicker/year-picker/year-picker.trans.styles.scss
