+++ libs/core/src/components/index.ts
@@ -39,6 +39,6 @@ export * from './textarea'
 export * from './radio'
 export * from './alert'
 export * from './rich-text'
-export * from './year-picker'
-export * from './month-picker'
+export * from './datepicker/year-picker'
+export * from './datepicker/month-picker'
 export * from './fab'
