+++ libs/core/src/components/pure.ts
@@ -40,5 +40,5 @@ export * from './radio/radio-group/radio-group.component'
 export * from './radio/radio.component'
 export * from './checkbox/checkbox-group/checkbox-group.component'
 export * from './checkbox/checkbox.component'
-export * from './year-picker/year-picker.component'
-export * from './month-picker/month-picker.component'
+export * from './datepicker/year-picker/year-picker.component'
+export * from './datepicker/month-picker/month-picker.component'
